revised ata1 and ata1 device identifier loop.
For some reason (I'm sure there is an explanation) /sys/*/sd* host bus identifiers changes fro 1:0:0:0 to 2:0:0:0 to 3:0:0:0. To avoid adding more and more else if statements, I revised the selection criteria - now if its on ata1 and either 0 or 1 or ata2 either 0 or 1 of the 2nd field (the third position above) it selects the appropriate bay.